requirement.txt
aioredis==1.2.0
amqp==5.3.1
asgiref==3.9.1
async-timeout==3.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
certifi==2024.8.30
channels==4.3.1
channels-redis==4.3.0
chardet==3.0.4
constantly==23.10.4
daphne==4.2.1
defusedxml==0.7.1
Django==5.2.5
django-extensions==4.1
django-allauth==65.11.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-progressbarupload==0.1.7
dj-rest-auth==7.0.1
django-rest-framework==0.1.0
django-timezone-field==7.1
djangorestframework==3.16.1
#djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.5.1
dnspython==2.7.0
drf-yasg==1.21.10
ffmpeg-python==0.1.17
future==0.17.1
hiredis==2.3.2
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
kafka-python==1.4.3
kombu==5.5.4
Markdown==3.0.1
msgpack==1.1.1
mysqlclient==2.2.7
oauthlib==3.3.1
paho-mqtt==2.1.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymongo==4.14.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-etcd==0.4.5
python3-openid==3.1.0
pytz==2024.2
redis==6.4.0
requests==2.32.3
requests-oauthlib==2.0.0
schedule==1.2.2
six==1.16.0
Twisted==25.5.0
txaio==25.6.1
urllib3==2.2.2
vine==5.1.0
zope.interface==7.2
